+++ accounts-db/src/lib.rs
@@ -55,9 +55,7 @@ pub mod waitable_condvar;
 
 pub use {
     buffered_reader::large_file_buf_reader,
-    file_io::{
-        file_creator, set_path_permissions, validate_memlock_limit_for_disk_io, FileCreator,
-    },
+    file_io::{file_creator, set_path_permissions, FileCreator},
     obsolete_accounts::{ObsoleteAccountItem, ObsoleteAccounts},
 };
